feat: New Color namespace to handle RGBA values passed in rendering functions
The addition of this namespace makes for a convienient place to get the
colors of MatrixTile, TetrominoType, and other colors used in the game.
This has also helped with cleaning up a lot of nasty code in
MatrixRenderer and Renderer. Now these functions, being free, can be
used anywhere in the codebase